xen, blkfront: prepare frontend to use multiple queues
This commit factors out ring-related fields from the blkfront_info
structure into a blkfront_ring_info structure, and prepares the
code to allocate and use multiple rings according to the number of
supported hardware queues.
As of now, the whole set of rings is protected by a single lock,
and grant-related fields are still in the blkfront_info structure.

NOTE: no mapping is performed yet between hardware context and ring;
      a mapping function will need to be added in the near future.

Also note: noisy commit; the only relevant part is probably in the
           factoring of data structures and in the device init logic.