xen, blkfront: implement mapping between hctx and ring
This commit adds a real mapping between hardware context and ring.
The mapping is performed by adding an init_hctx hook to the driver;
the hook stores in the driver_data field of the hardware context
a pointer to the ring associated to it, by using as a key for the
mapping the hardware context index and a simple 1:1 matching.
This commit also lets the hardware context index be stored in the
blkfront_ring_info structure, so that it can be successfully
retrieved when needed.
Note that the interrupt routines don't need to know anything about
the hardware context, as are initialized to carry the pointer to
the ring-related structure they are associated to.