xen, blkfront: gather the number of hw queues from the backend
This commit lets the frontend gather the number of available submit
queues from the backend. This also changed the previously-added
feature_multiqueue flag to keep the number of queues gathered from
the backend. The fake value of zero is kept in the variabile whether
the key couldn't be found to signal that the backend does not support
the feature.