drivers, block: add first version of fake multi-queue test driver
The driver added by this commit returns a fixed value for each of
the submit queues, so that consistency of access to hardware ctx
can be verified very easily. As of now, one only device is created
as soon as the module is loaded, and the number of hardware queues
equals the number of available CPUs.