xen, blkfront: add support for fallback XenStore keys
This commit adds to the frontend driver support for handling the
case in which the host driver does not support multi-queue; in this
case, either the blkback driver advertises nr_supported_hw_queues as 0
or the nr_supported_hw_queues key does not exist in the XenStore.
Fallback keys are used to advertise the only ring's ref and evtchn.