xen, blkfront: gather the number of hardware queues as soon as possible
This commit moves the code gathering the number of available hardware
queues to be executed on device probe. At that time, in fact, the
backend will already have advertised it, and the correct number of
rings can be allocated also in the frontend. The frontend can, in this
way, also advertise the correct information in the XenStore.

Note: not stress tested yet.