xen, blkback: add comment about retrieving ringrefs and evtchns
This commit adds a triple-X comment about the code retrieving
ring references and event channels. Even if the device supports
multiqueue, and the backend does, it is not sure that also the
guest's frontend supports multiqueue. Not being able to read
ring references and event channels with the new naming convention
does not therefore imply a failure. It will be necessary to add
code to fall back to searching for keys with the old names before
returning with an error.