block, fake_mq_blk: allow for variable number of submit queues
This commit adds to the fake_mq_blk driver a parameter that allows
the user to choose the number of submi queues as the driver is loaded.
If the number of submit queues is not specified, the driver keeps its
current behavior, that is assigning a submit queue to each online
CPU core. In case the chosen number of submit queue is not equal to
the number of online CPUs, the requests' buffer is not filled with
any value.