xen, blkback: allocate memory for multiple rings only once
Since the number of available submit queues is retrieved very soon,
e.g., before there is any need to queue a request, there is no need
to allocate a dummy ring. Also, re-allocating contiguous memory for
an array of large structures might let the start address of the
memory area to change, therefore provoking unpredictable issues.
This commit lets the memory for the array of rings be allocated
only once.