xen, blkfront: make fields related to grants and indirect descs per-ring
This commit is a first step of preparing the data structures of the
frontend driver to move locking to be per-ring. It begins by moving
fields related to persistent grants and indirect descriptors to be
per-ring, still not resizing the per-ring pool and keeping the lock
per-device. The callback executed when a grant has been freed is
also per-device.