xen, blkback: use two rings for requests and responses
This commit adapts the backend's data structures to factor out the
strictly ring-related fields and to arrange for a request and
a response ring. Locking is still kept intact, whereas following
commits will extend the visibility of some ring fields to allow
to separate the domain of handling the rings.