xen, blkfront: negotiate the number of block rings with the backend
This commit implements the negotiation of the number of block rings
to be used; as a default, the number of rings is decided by the
frontend driver and is equal to the number of hardware queues that
the backend makes available. In case of guest migration towards a
host whose devices expose a different number of hardware queues, the
number of I/O rings used by the frontend driver remains the same;
XenStore keys may vary if the frontend needs to be compatible with
a host not having multi-queue support.

Signed-off-by: Bob Liu <bob.liu@oracle.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>